requirements.txt
torch>=2.0.1
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
optuna>=3.3
jupyter>=7.0
pyyaml>=6.0
tqdm>=4.65